pyproject.toml
[project]
name = "avl-axi-stream"
version = "0.2.0"
description = "Apheleia Verification Library - AMBA AXI Stream verification component."
authors = [{name = "avl", email = "avl@projectapheleia.net"}]
maintainers = [{name = "avl", email = "avl@projectapheleia.net"}]
keyword = ["hdl", "verification", "AVL", "UVM", "AMBA", "AXI", "AXI-STREAM"]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
dependencies = [
    "matplotlib>=3.10.0",
    "avl-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=0.910",
    "black>=22.3.0",
    "sphinx>=8.1.3",
    "sphinx_rtd_theme>=3.0.2",
    "ruff>=0.11.0"
]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "avl_axi_stream" }]

[tool.ruff]
target-version = "py310"

src = ["avl", "examples", "bin"]

select = ["E", "F", "B", "I", "UP", "C90"]

exclude = ["doc", "build", "dist", "venv"]

line-length = 200

fix = false

[tool.ruff.mccabe]
max-complexity = 30

[project.scripts]
avl-axi-stream-get-verilog = "avl_axi_stream._tools:get_verilog"

[tool.setuptools.package-data]
avl-apb = ["rtl/*.v"]
